Burdukov 381908-1 lab2 horse min range

+++ modules/burdukov_mikhail_horse_min_range/src/horse_min_range.cpp
@@ -17,7 +17,7 @@ int minHorseRange::calc_range() const {
     auto x = q.front();
     q.pop();
     if (x == finish) break;
-    for (int i = 0; i < di.size(); i++) {
+    for (size_t i = 0; i < di.size(); i++) {
       auto y(x);
       y.first += di[i];
       y.second += dj[i];
